requirements.txt
numpy==1.22.4
dearpygui==1.6.2
opencv-contrib-python==4.6.0.66
onnxruntime==1.11.1
black==22.3.0
mediapipe==0.8.10
protobuf==3.20.1
azure-iot-device==2.11.0
